Correct task 3 not promoting float to doubles

+++ variadic_functions/3-print_all.c
@@ -34,7 +34,7 @@ void print_all(const char * const format, ...)
 				printf("%s%d", separator, va_arg(strings, int));
 				break;
 			case 'f':
-				printf("%s%d", separator, va_arg(strings, int));
+				printf("%s%f", separator, va_arg(strings, double));
 				break;
 			case 's':
 				string = va_arg(strings, char *);
